+++ src/Plugin.php
@@ -30,7 +30,7 @@ use yii\web\Response;
  */
 class Plugin extends BasePlugin
 {
-    public string $schemaVersion = '1.11.0';
+    public string $schemaVersion = '1.12.0';
     public bool $hasCpSettings = true;
     public bool $hasCpSection = true;